refactor: metrics chart to component
